Conditionally load Cordova and leverage a plugin.
The Cordova assets are loaded via a custom cdvfile URI scheme which must
be whitelisted. The whitelisting is done in the HTML content with a Content
Security Policy meta tag and also in the manifest.json file. The latter
is done so that ManifoldJS allows the request to go through.